% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_parm.R
\name{coerce_parm}
\alias{coerce_parm}
\alias{theta2K}
\alias{K2theta}
\alias{theta2K.rcon}
\alias{theta2K.rcor}
\alias{K2theta.rcon}
\alias{K2theta.rcor}
\title{Coerce parameters in RCOX models.}
\usage{
theta2K(object, theta, scale = "original")

K2theta(object, K, scale = "original")

\method{theta2K}{rcon}(object, theta, scale = "original")

\method{theta2K}{rcor}(object, theta, scale = "original")

\method{K2theta}{rcon}(object, K, scale = "original")

\method{K2theta}{rcor}(object, K, scale = "original")
}
\arguments{
\item{object}{Model object.}

\item{theta}{Parameter.}

\item{scale}{On which scales are the parameters computed.}

\item{K}{Concentration matrix.}
}
\description{
Coerce parameters in RCOX models. These functions
  should not be called directly by the user.
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
