% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grc_summary_doc.R
\name{grc-summary}
\alias{grc-summary}
\title{The package `gRc': summary information}
\description{
This package is for statistical inference in RCOX models. That is,
graphical Gaussian models where specific entries of the inverse
covariance matrix or partial correlation matrix have been
restricted to being equal. Entries which are restricted to being
identical are displayed with identical colours in the independence
graph. Hence the name of the package gRc: The "c" stands for
colours.
}
\details{
#' @author  Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\section{Details}{
 The function for specifying RCOX models is rcox, and we
refer to the help page for that function for examples.
}

\seealso{
\code{\link{rcox}}
}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
