\name{gRc}
\alias{gRc}
\title{
The package `gRc': summary information
}
\description{
  This package is for statistical inference in RCOX models. That is,
  graphical Gaussian models where specific entries of the inverse
  covariance matrix or partial correlation matrix have been restricted
  to being equal. Entries which are restricted to being identical are
  displayed with identical colours in the independence graph. Hence the
  name of the package gRc: The "c" stands for colours.
}
\section{Details}{
  The function for specifying RCOX models is rcox, and we refer to
  the help page for that function for examples.
}


\section{Authors}{
  Søren Højsgaard, sorenh@agrsci.dk
}
\seealso{\code{\link{rcox}}}

\keyword{multivariate}
\keyword{models}
\keyword{graphs}
