\name{internals}

%%% Utilities
%%% Code: utility.R
\alias{coef.rcox}
\alias{vcov.rcox}
\alias{summary.rcox}
\alias{print.summary.rcox}
\alias{tocc}
\alias{cc2str}
\alias{cholSolve}
\alias{print.coloutClass}


%%% Repair K
%%% Code: repairK.R
\alias{findKinModel}
\alias{findKinModel.rcox}


%%% Printing, plotting etc.
%%% code: All over the place
\alias{print.colourClass}
\alias{print.rcox}
\alias{plot.rcox}
\alias{print.statTable}
%\alias{}

%%% Symmetrize, make p.d. and similar matrix operations
%%% code: repairK.R
\alias{regularizeK}
%\alias{}

%%% calculateVCOV
%%% code: calculateVCOV.R
\alias{calculateVCOV}










\title{Internal functions for the gRc package}
\description{
  These functions should not be called directly by the user.
}

\author{Søren Højsgaard, sorenh@agrsci.dk}
\keyword{internal}% at least one, from doc/KEYWORDS

