% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{plot.rcox}
\alias{plot.rcox}
\title{Plot rcox object}
\usage{
\method{plot}{rcox}(x, y, ...)
}
\arguments{
\item{x}{An rcox object}

\item{y}{Ignored}

\item{...}{Currently not used}
}
\description{
Plot rcox object
}
\examples{
gm  = ~al:an:st
vcc = list(~me+st, ~ve+an, ~al)
ecc = list(~me:ve+me:al, ~ve:al+al:st)
m1 <- rcox(gm=gm, vcc=vcc, ecc=ecc, data=math, method='matching')
plot(m1)

}
