% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-op.R
\name{set-op}
\alias{set-op}
\alias{cardOrder}
\alias{listOrder}
\alias{listOrder.list}
\alias{listOrder.numeric}
\alias{listOrder.default}
\title{Set operations}
\description{
These functions are not intended to be called by the
  user.
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
