% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx2yyy.R
\name{xxx2yyy}
\alias{xxx2yyy}
\alias{formula2names}
\alias{formula2names.default}
\alias{formula2names.list}
\alias{formula2names.formula}
\alias{formula2names.NULL}
\alias{names2indices}
\alias{names2formula}
\alias{names2formula.list}
\alias{names2formula.default}
\alias{indices2vectors}
\alias{formula2string}
\alias{formula2string.list}
\alias{formula2string.default}
\alias{getIndex}
\alias{getIndex.default}
\alias{getIndex.list}
\alias{getIndex.character}
\alias{ecc2edges}
\alias{cc2formula}
\title{Conversions of the type xxx2yyy}
\description{
Conversions of the type xxx2yyy.
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
