% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_generating_class.R
\name{generating-class}
\alias{generating-class}
\title{Generating class for RCON / RCOR models}
\description{
Implementation of generating classes for RCOX
  models. These functions are not intended to be called by the
  user.
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
